reconfigure pokemon card style and details
