+++ frontend/src/components/PokemonCard.tsx
@@ -2,6 +2,8 @@ import { Card, CardContent, Typography, CardMedia, Chip, LinearProgress, Box } f
 import { usePokemonDetails } from '../hooks/usePokemonDetails'; // Asegúrate de ajustar la ruta de importación según tu estructura de directorios
 import { useState } from 'react';
 import PokemonDetailsModal from './PokemonDetails';
+import Image from 'next/image';
+import { getTypeColorClass } from '@/utils/getTypeColors';
 
 interface PokemonCardInterface {
     pokemonUrl: string;
@@ -23,40 +25,31 @@ const PokemonCard = ({ pokemonUrl }:PokemonCardInterface) => {
   if (isError) return <div>Error loading details.</div>;
   if (isLoading || !details) return <div>Loading details...</div>;
 
-  const normalize = (value:any) => value * 100 / 255; // Normalizo los valores para la barra de progreso.
-
 
   return (
     <>
-      <Card onClick={handleClickOpen} className=" max-h-[100%] cursor-pointer transition duration-300 ease-in-out transform hover:-translate-y-1 hover:scale-90 bg-white shadow-lg rounded-lg overflow-hidden">
-        <CardMedia
-          component="img"
-          image={`https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/${details.id}.png`}
-          alt={details.name}
-          className="w-full h-50 object-fit"
-          />
-        <CardContent>
-          <Typography gutterBottom variant="h4" component="div" className="text-center capitalize">
-            {details.name}
-          </Typography>
-          <div className="flex justify-center flex-wrap gap-2 mb-4">
-            {details.types.map((type:any, index:any) => (
-              <Chip key={index} label={type.type.name} className="capitalize" color="primary" />
-            ))}
-          </div>
-          <div className="space-y-2">
-            {details.stats.map((stat:any, index:any) => (
-              <Box key={index}>
-                  <Typography variant="body1" color="text.secondary" className="font-medium">
-                    {stat.stat.name.replace('-', ' ')}: {stat.base_stat}
-                  </Typography>
-                  <LinearProgress variant="determinate" value={normalize(stat.base_stat)} />
-              </Box>
-            ))}
-          </div>
-        </CardContent>
-      </Card>
-      <PokemonDetailsModal open={open} handleClose={handleClose} details={details} />
+    <div className=''>
+			<div onClick={handleClickOpen} className='cursor-pointer hover:-translate-y-2 hover:scale-110 hover:bg-indigo-500 duration-300 bg-gray-200 flex items-center justify-center rounded-xl h-[250px]'>
+				<Image
+					src={details.sprites.other.dream_world.front_default}
+					alt={`Pokemon ${details.name}`}
+          width={200}
+          height={100}
+				/>
+			</div>
+			<div className='p-5'>
+				<Typography className='block text-black mb-5'>N° {details.id}</Typography>
+				<Typography variant="h4"><strong>{details.name}</strong></Typography>
+				<div className='flex gap-2 border-r-3 text-black'>
+					{details.types.map((type:any) => (
+						<Typography variant='h6'  key={type.type.name} className={`${getTypeColorClass(type.type.name)} mt-2 px-4 py-1 rounded-xl   `}>
+							{type.type.name}
+						</Typography>
+					))}
+				</div>
+			</div>
+		</div>
+    <PokemonDetailsModal open={open} handleClose={handleClose} details={details} />
     </>
   );
 };
