+++ frontend/src/components/PokemonDetails.tsx
@@ -1,7 +1,9 @@
-import { Dialog, DialogTitle, DialogContent, DialogActions, Button, Typography, Chip, Grid, Card, CardMedia, Box } from '@mui/material';
+import { Dialog, DialogTitle, DialogContent, DialogActions, Button, Typography, Chip, Grid, Card, CardMedia, Box, LinearProgress } from '@mui/material';
 
 const PokemonDetailsModal = (props:any) => {
     const {open, handleClose, details} = props;
+
+    const normalize = (value:any) => value * 100 / 255; // Normalizo los valores para la barra de progreso.
     return (
       <Dialog open={open} onClose={handleClose} fullWidth maxWidth="sm">
         <Box className="flex flex-col items-center ">
@@ -19,14 +21,17 @@ const PokemonDetailsModal = (props:any) => {
           </Grid>
           <Grid item xs={12} sm={6} >
             <Box className="flex flex-col justify-start items-start gap-3">
-            <Typography variant="body1"><strong>Height:</strong> {details.height} dm</Typography>
-            <Typography variant="body1"><strong>Weight:</strong> {details.weight} hg</Typography>
-            <Typography variant="body1"><strong>Types:</strong>
-           
-              {details.types.map((type:any, index:any) => (
-                  <Chip key={index} label={type.type.name} style={{ margin: 2 }} />
-                ))}
-            </Typography>
+            <div className="space-y-2">
+            <Typography variant="body1"><strong>Stats:</strong></Typography>
+            {details.stats.map((stat:any, index:any) => (
+              <Box key={index}>
+                  <Typography variant="body1" color="text.secondary" className="font-medium">
+                    {stat.stat.name.replace('-', ' ')}: {stat.base_stat}
+                  </Typography>
+                  <LinearProgress variant="determinate" value={normalize(stat.base_stat)} />
+              </Box>
+            ))}
+          </div>
             <Typography variant="body1"><strong>Abilities:</strong>
               {details.abilities.map((ability:any, index:any) => (
                   <Chip key={index} label={ability.ability.name} style={{ margin: 2 }} />
